Release only on tags
The idea of a "Nightly" for Github does not work, as GH releases must be
tagged and it is highly discouraged to change a tag after it was
published.
Hence there are now only releases on (manual) tags. Those should be of
the form 'v1.2.3' or 'v1.2.3-rc4' as per semantic versioning.