Fix handling of text archives (GER/ENG)
The ENG files have a list of C-Strings (i.e. NULL terminated)
Duplicated strings are saved only once but referenced multiple times (see NAMES.ENG)
This means that the length calculation by using the next element will fail (actually the check before catches that and refuses to load the file)
Also empty strings are stored (as a single NULL byte), so there is a difference between en empty element and an element with an empty string

This is now correctly implemented in saving and loading.